Use gradle-plugin-test-utils library for tests
