+++ build.gradle.kts
@@ -45,6 +45,8 @@ dependencies {
 
     testImplementation("com.squareup.okhttp3:mockwebserver:4.9.0")
 
+    testImplementation("org.unbroken-dome.gradle-plugin-utils:gradle-plugin-test-utils:0.1.0")
+
     testImplementation(kotlin("stdlib-jdk8"))
     testImplementation(kotlin("reflect"))
 }
