--- src/test/kotlin/org/unbrokendome/gradle/plugins/helm/GradleTaskTestContext.kt
@@ -1,10 +1,0 @@
-package org.unbrokendome.gradle.plugins.helm
-
-import org.gradle.api.Task
-
-
-class GradleTaskTestContext<T : Task>(
-    val taskType: Class<T>,
-    val defaultTaskName: String = taskType.simpleName.decapitalize(),
-    val defaultConfig: T.() -> Unit = {}
-)
