+++ src/test/kotlin/org/unbrokendome/gradle/plugins/helm/release/HelmReleasesPluginTest.kt
@@ -19,20 +19,20 @@ import org.unbrokendome.gradle.plugins.helm.release.dsl.HelmReleaseTarget
 import org.unbrokendome.gradle.plugins.helm.release.dsl.activeReleaseTarget
 import org.unbrokendome.gradle.plugins.helm.release.dsl.releaseTargets
 import org.unbrokendome.gradle.plugins.helm.release.dsl.releases
-import org.unbrokendome.gradle.plugins.helm.spek.applyPlugin
 import org.unbrokendome.gradle.plugins.helm.spek.propertyMappingInfo
 import org.unbrokendome.gradle.plugins.helm.spek.propertyMappingTests
-import org.unbrokendome.gradle.plugins.helm.spek.setupGradleProject
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.asType
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.containsTask
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.doesNotContainItem
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.doesNotHaveTaskDependency
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.hasExtension
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.hasExtensionNamed
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.hasOnlyTaskDependency
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.hasTaskDependency
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.hasValueEqualTo
-import org.unbrokendome.gradle.plugins.helm.testutil.evaluate
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.cast
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.containsTask
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.doesNotContainItem
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.doesNotHaveTaskDependency
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.hasExtension
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.hasExtensionNamed
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.hasOnlyTaskDependency
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.hasTaskDependency
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.hasValueEqualTo
+import org.unbrokendome.gradle.pluginutils.test.evaluate
+import org.unbrokendome.gradle.pluginutils.test.spek.applyPlugin
+import org.unbrokendome.gradle.pluginutils.test.spek.setupGradleProject
 import java.net.URI
 import java.time.Duration
 
@@ -290,7 +290,7 @@ object HelmReleasesPluginTest : Spek({
 
             assertThat(project.helm, name = "helm")
                 .hasExtensionNamed("activeReleaseTarget")
-                .asType<Property<String>>()
+                .cast<Property<String>>()
                 .hasValueEqualTo("test-target")
         }
     }
