+++ src/test/kotlin/org/unbrokendome/gradle/plugins/helm/spek/PropertyMappingTest.kt
@@ -9,9 +9,9 @@ import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
 import org.spekframework.spek2.style.specification.Suite
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.containsTask
-import org.unbrokendome.gradle.plugins.helm.testutil.assertions.hasValueEqualTo
 import org.unbrokendome.gradle.pluginutils.asFile
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.containsTask
+import org.unbrokendome.gradle.pluginutils.test.assertions.assertk.hasValueEqualTo
 import java.io.File
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty1
