--- src/test/kotlin/org/unbrokendome/gradle/plugins/helm/testutil/assertions/Assertions.kt
@@ -1,9 +1,0 @@
-package org.unbrokendome.gradle.plugins.helm.testutil.assertions
-
-import assertk.Assert
-
-
-fun <T> Assert<*>.asType(): Assert<T> = transform { actual ->
-    @Suppress("UNCHECKED_CAST")
-    actual as T
-}
