--- src/test/kotlin/org/unbrokendome/gradle/plugins/helm/testutil/assertions/GradleProject.kt
@@ -1,24 +1,0 @@
-package org.unbrokendome.gradle.plugins.helm.testutil.assertions
-
-import assertk.Assert
-import assertk.assertions.isInstanceOf
-import assertk.assertions.support.expected
-import org.gradle.api.Project
-import org.gradle.api.Task
-import kotlin.reflect.KClass
-
-
-private fun Assert<Project>.containsTaskInternal(taskName: String) =
-    transform("task \"$taskName\"") { actual ->
-        actual.tasks.findByName(taskName)
-            ?: expected("to contain a task named \"$taskName\"")
-    }
-
-
-fun <T : Task> Assert<Project>.containsTask(taskName: String, taskType: KClass<T>) =
-    containsTaskInternal(taskName)
-        .isInstanceOf(taskType)
-
-
-inline fun <reified T : Task> Assert<Project>.containsTask(taskName: String) =
-    containsTask(taskName, T::class)
